chore: reorganize test files
