+++ package.json
@@ -65,7 +65,7 @@
     "modulePathIgnorePatterns":[
       "src/globalConfig.json"
     ],
-    "rootDir": "src",
+    "rootDir": "test",
     "testRegex": ".*\\.spec\\.ts$",
     "transform": {
       "^.+\\.(t|j)s$": "ts-jest"
